refactor: move Sanitizer to internals
