+++ clients/src/main/java/org/apache/kafka/common/utils/AppInfoParser.java
@@ -20,6 +20,7 @@ import org.apache.kafka.common.MetricName;
 import org.apache.kafka.common.metrics.Gauge;
 import org.apache.kafka.common.metrics.Metrics;
 
+import org.apache.kafka.common.utils.internals.Sanitizer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
